fix: file path and group loading problem
