+++ package.json
@@ -53,6 +53,7 @@
     "eslint-config-prettier": "^10.1.2",
     "eslint-formatter-visualstudio": "^8.40.0",
     "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
     "globals": "^16.0.0",
     "husky": "^9.1.7",
     "lint-staged": "^15.5.1",
