+++ src/hooks/useGetGroupById.ts
@@ -1,7 +1,7 @@
 // src/api/queries/useGetGroupById.ts
 
 import { useQuery } from "@tanstack/react-query";
-import { getGroupById } from "@/api/groups/getGroupById";
+import { getGroupById } from "@/lib/request/getGroupById";
 
 export function useGetGroupById(id: string | undefined) {
   return useQuery({
