+++ src/hooks/useRemoveMember.ts
@@ -1,23 +1,21 @@
-// src/api/mutations/useAddMember.ts
+// src/api/mutations/useRemoveMember.ts
 
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { addMember } from "@/api/groups/addMember";
+import { removeMember } from "@/lib/request/removeMember";
 
-type AddMemberRequest = {
-  member: string;
-  role: string;
-};
-
-type AddMemberOptions = {
+type RemoveMemberOptions = {
   onSuccess?: () => void | Promise<void>;
   onError?: (err: unknown) => void;
 };
 
-export function useAddMember(groupId: string, options?: AddMemberOptions) {
+export function useRemoveMember(
+  groupId: string,
+  options?: RemoveMemberOptions,
+) {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: (members: AddMemberRequest[]) => addMember(groupId, members),
+    mutationFn: (memberId: string) => removeMember(groupId, memberId),
 
     onSuccess: async () => {
       await queryClient.invalidateQueries({ queryKey: ["members", groupId] });
